backend/requirements.txt
fastapi==0.111.*
uvicorn[standard]==0.29.*
strands-agents>=1.0.0
httpx==0.27.*
pydantic==2.7.*
jinja2==3.1.*
weasyprint==62.*
python-dotenv==1.0.*
boto3 